completely refactored HttpAPI into more manageable routes.
