+++ app/library/HttpAPI.py
@@ -1,140 +1,45 @@
-import asyncio
 import base64
-import functools
 import hmac
-import json
+import importlib
+import inspect
 import logging
-import random
-import time
-import uuid
+import pkgutil
 from collections.abc import Awaitable
 from datetime import UTC, datetime, timedelta
 from pathlib import Path
-from urllib.parse import unquote_plus, urlparse
+from typing import Any
 
 import anyio
-import httpx
-import magic
 from aiohttp import web
 from aiohttp.web import Request, RequestHandler, Response
-from library.ag_utils import ag
-from yt_dlp.cookies import LenientSimpleCookie
 
 from .cache import Cache
-from .common import Common
-from .conditions import Condition, Conditions
 from .config import Config
 from .DownloadQueue import DownloadQueue
 from .encoder import Encoder
-from .Events import EventBus, Events, message
+from .Events import EventBus
 from .ffprobe import ffprobe
-from .ItemDTO import Item
-from .M3u8 import M3u8
-from .Notifications import Notification, NotificationEvents
-from .Playlist import Playlist
-from .Presets import Preset, Presets
-from .Segments import Segments
-from .Subtitle import Subtitle
-from .Tasks import Task, Tasks
-from .Utils import (
-    REMOVE_KEYS,
-    StreamingError,
-    arg_converter,
-    decrypt_data,
-    encrypt_data,
-    extract_info,
-    get_file,
-    get_file_sidecar,
-    get_files,
-    get_mime_type,
-    init_class,
-    read_logfile,
-    validate_url,
-    validate_uuid,
-)
-from .YTDLPOpts import YTDLPOpts
+from .router import get_routes
+from .Utils import decrypt_data, encrypt_data, get_file, get_mime_type
 
-LOG = logging.getLogger("http_api")
-MIME = magic.Magic(mime=True)
+LOG: logging.Logger = logging.getLogger("http_api")
 
 
-class HttpAPI(Common):
-    _static_holder: dict = {}
-    """Holds loaded static assets."""
+class HttpAPI:
+    def __init__(self, root_path: str, queue: DownloadQueue):
+        self.queue: DownloadQueue = queue or DownloadQueue.get_instance()
+        self.encoder: Encoder = Encoder()
+        self.config: Config = Config.get_instance()
+        self._notify: EventBus = EventBus.get_instance()
 
-    _ext_to_mime: dict = {
-        ".html": "text/html",
-        ".css": "text/css",
-        ".js": "application/javascript",
-        ".json": "application/json",
-        ".ico": "image/x-icon",
-    }
-    """Map ext to mimetype"""
-
-    _frontend_routes: list[str] = [
-        "/console",
-        "/presets",
-        "/tasks",
-        "/notifications",
-        "/changelog",
-        "/logs",
-        "/conditions",
-        "/browser",
-        "/browser/{path:.*}",
-    ]
-    """Frontend routes to be preloaded"""
-
-    def __init__(
-        self,
-        root_path: str,
-        queue: DownloadQueue | None = None,
-        encoder: Encoder | None = None,
-        config: Config | None = None,
-    ):
-        self.queue = queue or DownloadQueue.get_instance()
-        self.encoder = encoder or Encoder()
-        self.config = config or Config.get_instance()
-        self._notify = EventBus.get_instance()
-
-        self.rootPath = root_path
-        self.routes = web.RouteTableDef()
+        self.rootPath: str = root_path
         self.cache = Cache()
         self.app: web.Application | None = None
-        self.isRequestingBackground = False
-
-        super().__init__(queue=self.queue, encoder=self.encoder, config=self.config)
-
-    @staticmethod
-    def route(method: str, path: str, name: str | None = None) -> Awaitable:
-        """
-        Decorator to mark a method as an HTTP route handler.
-
-        Args:
-            method (str): The HTTP method.
-            path (str): The path to the route.
-            name (str | None): The name of the route (optional).
-
-        Returns:
-            Awaitable: The decorated function.
-
-        """
-
-        def decorator(func):
-            @functools.wraps(func)
-            async def wrapper(*args, **kwargs):
-                return await func(*args, **kwargs)
-
-            wrapper._http_method = method.upper()
-            wrapper._http_path = path
-            wrapper._http_name = name
-            return wrapper
-
-        return decorator
 
     async def on_shutdown(self, _: web.Application):
         pass
 
-    def attach(self, app: web.Application) -> "HttpAPI":
+    def attach(self, app: web.Application):
         """
         Attach the routes to the application.
 
@@ -152,6 +57,7 @@ class HttpAPI(Common):
                 app=app,
                 base_path=self.config.base_path.rstrip("/"),
                 download_path=self.config.download_path,
+                this=self,
             )
         )
 
@@ -176,114 +82,8 @@ class HttpAPI(Common):
             LOG.exception(e)
 
         app.on_shutdown.append(self.on_shutdown)
-        return self
 
-    async def _static_file(self, req: Request) -> Response:
-        """
-        Preload static files from the ui/exported folder.
-
-        Args:
-            req (Request): The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        path = req.path
-
-        if "/" != self.config.base_path and path.rstrip("/") == self.config.base_path.rstrip("/"):
-            path: str = f"{self.config.base_path.rstrip('/')}/index.html"
-
-        if path not in self._static_holder:
-            for k in self._static_holder:
-                if path.startswith(k):
-                    path = k
-                    break
-            else:
-                return web.json_response(
-                    {"error": "File not found.", "file": path}, status=web.HTTPNotFound.status_code
-                )
-
-        item: dict = self._static_holder[path]
-
-        return web.Response(
-            body=item.get("content"),
-            headers={
-                "Pragma": "public",
-                "Cache-Control": "public, max-age=31536000",
-                "Content-Type": item.get("content_type"),
-                "X-Via": "memory" if not item.get("file") else "disk",
-            },
-            status=web.HTTPOk.status_code,
-        )
-
-    def _preload_static(self, app: web.Application) -> "HttpAPI":
-        """
-        Preload static files from the ui/exported folder.
-
-        Args:
-            app (web.Application): The application to attach the routes to.
-
-        Returns:
-            HttpAPI: The instance of the HttpAPI.
-
-        """
-        staticDir = (Path(self.rootPath) / "ui" / "exported").absolute()
-        if not staticDir.exists():
-            staticDir = (Path(self.rootPath).parent / "ui" / "exported").absolute()
-            if not staticDir.exists():
-                msg = f"Could not find the frontend UI static assets. '{staticDir}'."
-                raise ValueError(msg)
-
-        preloaded = 0
-
-        base_path: str = str(Path(self.config.base_path).as_posix()).rstrip("/")
-
-        for file in staticDir.rglob("*.*"):
-            if ".map" == file.suffix:
-                continue
-
-            urlPath: str = f"{base_path}/{str(file.as_posix()).replace(f'{staticDir.as_posix()!s}/', '')}"
-
-            with open(file, "rb") as f:
-                content = f.read()
-
-            contentType = self._ext_to_mime.get(file.suffix, MIME.from_file(file))
-
-            self._static_holder[urlPath] = {"content": content, "content_type": contentType}
-            LOG.debug(f"Preloading static '{urlPath}'.")
-            app.router.add_get(urlPath, self._static_file)
-            preloaded += 1
-
-        if "/index.html" in self._static_holder:
-            for path in self._frontend_routes:
-                path: str = f"{base_path}/{path.lstrip('/')}"
-                self._static_holder[path] = self._static_holder["/index.html"]
-                app.router.add_get(path, self._static_file)
-                if "{" not in path:
-                    self._static_holder[path + "/"] = self._static_holder["/index.html"]
-                    app.router.add_get(path + "/", self._static_file)
-                LOG.debug(f"Preloading static route '{path}'.")
-                preloaded += 1
-
-        if "/" != self.config.base_path:
-            LOG.debug(f"adding base_path folder '{base_path}' to routes.")
-            app.router.add_get(base_path, self._static_file, name="_base_path")
-            app.router.add_get(f"{base_path}/", self._static_file, name="_base_path_slash")
-
-        if preloaded < 1:
-            message = f"Failed to find any static files in '{staticDir}'."
-            if self.config.ignore_ui:
-                LOG.warning(message)
-                return self
-
-            raise ValueError(message)
-
-        LOG.info(f"Preloaded '{preloaded}' static files.")
-
-        return self
-
-    def add_routes(self, app: web.Application) -> "HttpAPI":
+    def add_routes(self, app: web.Application):
         """
         Add the routes to the application.
 
@@ -297,43 +97,34 @@ class HttpAPI(Common):
         registered_options: list = []
 
         base_path: str = self.config.base_path.rstrip("/")
+        from app.routes._static import preload_static
 
-        for attr_name in dir(self):
-            method = getattr(self, attr_name)
-            if hasattr(method, "_http_method") and hasattr(method, "_http_path"):
-                http_path = method._http_path
-                if http_path.startswith("/"):
-                    http_path = method._http_path[1:]
+        self._load_modules(Path(self.rootPath) / "routes")
+        preload_static(self.rootPath, self.config)
 
-                opts = {}
-                if hasattr(method, "_http_name") and method._http_name:
-                    opts["name"] = method._http_name
+        async def options_handler(_: Request) -> Response:
+            return web.Response(status=204)
 
-                LOG.debug(f"Adding API route {method._http_method} {base_path}/{http_path}' {opts}.")
-                self.routes.route(method._http_method, f"{base_path}/{http_path}", **opts)(method)
+        for route in get_routes().values():
+            routePath: str = f"/{route.path.lstrip('/')}"
 
-                if http_path in registered_options:
-                    continue
+            if self.config.base_path == route.path:
+                pass
+            elif "" == base_path or not routePath.rstrip("/").startswith(base_path.rstrip("/")):
+                LOG.debug(f"Route path '{routePath}' does not start with base path '{base_path}'. Skipping.")
+                route.path = f"{base_path}/{route.path.lstrip('/')}"
 
-                async def options_handler(_: Request) -> Response:
-                    return web.Response(status=204)
+            LOG.debug(f"Add ({route.name}) {route.method}: {route.path}.")
 
-                if "name" in opts:
-                    opts["name"] = f"{opts['name']}_options"
+            app.router.add_route(route.method, route.path, handler=route.handler, name=route.name)
 
-                self.routes.route("OPTIONS", f"{base_path}/{http_path}", **opts)(options_handler)
-                registered_options.append(http_path)
+            if route.path in registered_options:
+                continue
 
-        self.routes.static(f"{base_path}/api/download/", self.config.download_path, name="download_static")
-        self._preload_static(app)
+            app.router.add_route("OPTIONS", route.path, handler=options_handler, name=f"{route.name}_opts")
+            registered_options.append(route.path)
 
-        try:
-            app.add_routes(self.routes)
-        except ValueError as e:
-            if "ui/exported" in str(e):
-                msg = f"Could not find the frontend UI static assets. '{e}'."
-                raise RuntimeError(msg) from e
-            raise
+        app.router.add_static(f"{base_path}/api/download/", self.config.download_path, name="download_static")
 
     @staticmethod
     def basic_auth(username: str, password: str) -> Awaitable:
@@ -422,7 +213,7 @@ class HttpAPI(Common):
         return middleware_handler
 
     @staticmethod
-    def middle_wares(app: web.Application, base_path: str, download_path: str) -> Awaitable:
+    def middle_wares(app: web.Application, base_path: str, download_path: str, this: "HttpAPI") -> Awaitable:
         @web.middleware
         async def middleware_handler(request: Request, handler: RequestHandler) -> Response:
             static_path = str(app.router["download_static"].url_for(filename=""))
@@ -443,8 +234,32 @@ class HttpAPI(Common):
                         },
                     )
 
+            kwargs: dict[str, Any] = {
+                "request": request,
+                "queue": this.queue,
+                "encoder": this.encoder,
+                "config": this.config,
+                "notify": this._notify,
+                "cache": this.cache,
+                "app": app,
+                "http_api": this,
+            }
+
             try:
-                response = await handler(request)
+                sig = inspect.signature(handler)
+                expected_args = sig.parameters.keys()
+
+                try:
+                    if 1 == len(expected_args) and "request" in expected_args:
+                        response = await handler(request)
+                    else:
+                        filtered = {k: v for k, v in kwargs.items() if k in expected_args}
+                        response = await handler(**filtered)
+                except TypeError as te:
+                    if "missing 1 required positional argument" in str(te) and "request" in str(te):
+                        response = await handler(request)
+                    else:
+                        raise
             except web.HTTPException as e:
                 return web.json_response(data={"error": str(e)}, status=e.status_code)
             except Exception as e:
@@ -457,15 +272,21 @@ class HttpAPI(Common):
             contentType: str | None = response.headers.get("content-type", None)
             if contentType and "/" != base_path and contentType.startswith("text/html"):
                 rewrite_path: str = base_path.rstrip("/")
-                content: str = (
-                    response.body.decode("utf-8")
-                    .replace('<base href="/">', f'<base href="{rewrite_path}/">')
-                    .replace('baseURL:""', f'baseURL:"{rewrite_path}/"')
+                async with await anyio.open_file(response._path, "rb") as f:
+                    content = await f.read()
+                    content: str = (
+                        content.decode("utf-8")
+                        .replace('<base href="/">', f'<base href="{rewrite_path}/">')
+                        .replace('baseURL:""', f'baseURL:"{rewrite_path}/"')
+                    )
+
+                return web.Response(
+                    body=content.encode("utf-8"),
+                    status=response.status,
+                    headers=response.headers,
                 )
 
-                response.body = content.encode("utf-8")
-
-            if isinstance(response, web.FileResponse):
+            if request.path.startswith(static_path) and isinstance(response, web.FileResponse):
                 try:
                     ff_info = await ffprobe(response._path)
                     mime_type = get_mime_type(ff_info.get("metadata", {}), response._path)
@@ -477,1607 +298,15 @@ class HttpAPI(Common):
 
         return middleware_handler
 
-    @route("GET", "api/ping", "ping")
-    async def ping(self, _: Request) -> Response:
-        """
-        Ping the server.
-
-        Args:
-            _: The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        await self.queue.test()
-        return web.json_response(data={"status": "pong"}, status=web.HTTPOk.status_code)
-
-    @route("POST", "api/yt-dlp/convert", "yt_dlp_convert")
-    async def yt_dlp_convert(self, request: Request) -> Response:
-        """
-        Convert the yt-dlp args to a dict.
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        post = await request.json()
-        args: str | None = post.get("args")
-
-        if not args:
-            return web.json_response(data={"error": "args param is required."}, status=web.HTTPBadRequest.status_code)
-
-        try:
-            response = {"opts": {}, "output_template": None, "download_path": None}
-
-            data = arg_converter(args, dumps=True)
-
-            if "outtmpl" in data and "default" in data["outtmpl"]:
-                response["output_template"] = data["outtmpl"]["default"]
-
-            if "paths" in data and "home" in data["paths"]:
-                response["download_path"] = data["paths"]["home"]
-
-            if "format" in data:
-                response["format"] = data["format"]
-
-            bad_options = {k: v for d in REMOVE_KEYS for k, v in d.items()}
-            removed_options = []
-
-            for key in data:
-                if key in bad_options.items():
-                    removed_options.append(bad_options[key])
-                    continue
-                if not key.startswith("_"):
-                    response["opts"][key] = data[key]
-
-            if len(removed_options) > 0:
-                response["removed_options"] = removed_options
-
-            return web.json_response(data=response, status=web.HTTPOk.status_code)
-        except Exception as e:
-            err = str(e).strip()
-            err = err.split("\n")[-1] if "\n" in err else err
-            err = err.replace("main.py: error: ", "").strip().capitalize()
-            return web.json_response(
-                data={"error": f"Failed to parse command options for yt-dlp. '{err}'."},
-                status=web.HTTPBadRequest.status_code,
-            )
-
-    @route("GET", "api/yt-dlp/url/info", "get_info")
-    async def get_info(self, request: Request) -> Response:
-        """
-        Get the video info.
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object
-
-        """
-        url = request.query.get("url")
-        if not url:
-            return web.json_response(data={"error": "URL is required."}, status=web.HTTPBadRequest.status_code)
-
-        try:
-            validate_url(url)
-        except ValueError as e:
-            return web.json_response(data={"error": str(e)}, status=web.HTTPBadRequest.status_code)
-
-        preset = request.query.get("preset")
-        if preset:
-            exists = Presets.get_instance().get(preset)
-            if not exists:
-                return web.json_response(
-                    data={"status": False, "message": f"Preset '{preset}' does not exist."},
-                    status=web.HTTPBadRequest.status_code,
-                )
-        else:
-            preset = self.config.default_preset
-
-        try:
-            key = self.cache.hash(url + str(preset))
-
-            if self.cache.has(key) and not request.query.get("force", False):
-                data = self.cache.get(key)
-                data["_cached"] = {
-                    "status": "hit",
-                    "key": key,
-                    "ttl": data.get("_cached", {}).get("ttl", 300),
-                    "ttl_left": data.get("_cached", {}).get("expires", time.time() + 300) - time.time(),
-                    "expires": data.get("_cached", {}).get("expires", time.time() + 300),
-                }
-                return web.Response(body=json.dumps(data, indent=4), status=web.HTTPOk.status_code)
-
-            opts = {}
-
-            if ytdlp_proxy := self.config.get_ytdlp_args().get("proxy", None):
-                opts["proxy"] = ytdlp_proxy
-
-            ytdlp_opts = YTDLPOpts.get_instance().preset(name=preset).add(opts).get_all()
-
-            data = extract_info(
-                config=ytdlp_opts,
-                url=url,
-                debug=False,
-                no_archive=True,
-                follow_redirect=True,
-                sanitize_info=True,
-            )
-
-            if "formats" in data:
-                for index, item in enumerate(data["formats"]):
-                    if "cookies" in item and len(item["cookies"]) > 0:
-                        cookies = [f"{c.key}={c.value}" for c in LenientSimpleCookie(item["cookies"]).values()]
-                        if len(cookies) > 0:
-                            data["formats"][index]["h_cookies"] = "; ".join(cookies)
-                            data["formats"][index]["h_cookies"] = data["formats"][index]["h_cookies"].strip()
-
-            data["_cached"] = {
-                "status": "miss",
-                "key": key,
-                "ttl": 300,
-                "ttl_left": 300,
-                "expires": time.time() + 300,
-            }
-
-            self.cache.set(key=key, value=data, ttl=300)
-
-            return web.Response(body=json.dumps(data, indent=4), status=web.HTTPOk.status_code)
-        except Exception as e:
-            LOG.exception(e)
-            LOG.error(f"Error encountered while getting video info for '{url}'. '{e!s}'.")
-            return web.json_response(
-                data={
-                    "error": "failed to get video info.",
-                    "message": str(e),
-                    "formats": [],
-                },
-                status=web.HTTPInternalServerError.status_code,
-            )
-
-    @route("GET", "api/yt-dlp/archive/recheck", "archive_recheck")
-    async def archive_recheck(self, _) -> Response:
-        """
-        Recheck the manual archive entries.
-
-        Args:
-            _ (Request): The request object.
-
-        Returns:
-            Response: The response object
-
-        """
-        manual_archive = self.config.manual_archive
-        if not manual_archive:
-            return web.json_response(
-                data={"error": "Manual archive is not enabled."}, status=web.HTTPNotFound.status_code
-            )
-
-        manual_archive = Path(manual_archive)
-        if not manual_archive.exists():
-            return web.json_response(
-                data={"error": "Manual archive file not found.", "file": manual_archive},
-                status=web.HTTPNotFound.status_code,
-            )
-
-        tasks = []
-        response = []
-
-        def info_wrapper(id: str, url: str) -> tuple[str, dict]:
-            try:
-                return (
-                    id,
-                    extract_info(
-                        config={
-                            "proxy": self.config.get_ytdlp_args().get("proxy", None),
-                            "simulate": True,
-                            "dump_single_json": True,
-                        },
-                        url=url,
-                        no_archive=True,
-                    ),
-                )
-            except Exception as e:
-                return (id, {"error": str(e)})
-
-        async with await anyio.open_file(manual_archive) as f:
-            # line format is "youtube ID - at: ISO8601"
-            async for line in f:
-                line = line.strip()
-
-                if not line or not line.startswith("youtube"):
-                    continue
-
-                id = line.split(" ")[1].strip()
-
-                if not id:
-                    continue
-
-                url = f"https://www.youtube.com/watch?v={id}"
-                key = self.cache.hash(id)
-
-                if self.cache.has(key):
-                    data = self.cache.get(key)
-                    response.append({id: bool(data.get("id", None)) if isinstance(data, dict) else False})
-                    continue
-
-                tasks.append(
-                    asyncio.get_event_loop().run_in_executor(None, lambda i=id, url=url: info_wrapper(id=i, url=url))
-                )
-
-        if len(tasks) > 0:
-            results = await asyncio.gather(*tasks)
-            for data in results:
-                if not data:
-                    continue
-
-                id, info = data
-                self.cache.set(key=self.cache.hash(id), value=info, ttl=3600 * 6)
-                response.append({id: bool(data.get("id", None)) if isinstance(data, dict) else False})
-
-        return web.json_response(data=response, status=web.HTTPOk.status_code, dumps=self.encoder.encode)
-
-    @route("GET", "api/history/add", "quick_add")
-    async def quick_add(self, request: Request) -> Response:
-        """
-        Add a URL to the download queue.
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object
-
-        """
-        url: str | None = request.query.get("url")
-        if not url:
-            return web.json_response(data={"error": "url param is required."}, status=web.HTTPBadRequest.status_code)
-
-        data = {
-            "url": url,
-        }
-
-        preset = request.query.get("preset")
-        if preset:
-            exists = Presets.get_instance().get(preset)
-            if not exists:
-                return web.json_response(
-                    data={"status": False, "message": f"Preset '{preset}' does not exist."},
-                    status=web.HTTPBadRequest.status_code,
-                )
-            data["preset"] = preset
-
-        try:
-            status = await self.add(item=Item.format(data))
-        except ValueError as e:
-            return web.json_response(data={"status": False, "message": str(e)}, status=web.HTTPBadRequest.status_code)
-
-        return web.json_response(
-            data={"status": status.get("status") == "ok", "message": status.get("msg", "URL added")},
-            status=web.HTTPOk.status_code,
-            dumps=self.encoder.encode,
-        )
-
-    @route("POST", "api/history", "history_item_add")
-    async def history_item_add(self, request: Request) -> Response:
-        """
-        Add a URL to the download queue.
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        data = await request.json()
-
-        if isinstance(data, dict):
-            data = [data]
-
-        items = []
-        for item in data:
-            try:
-                items.append(Item.format(item))
-            except ValueError as e:
-                return web.json_response(data={"error": str(e), "data": item}, status=web.HTTPBadRequest.status_code)
-
-        status = await asyncio.wait_for(
-            fut=asyncio.gather(*[self.add(item=item) for item in items]),
-            timeout=None,
-        )
-        response = []
-        for i, item in enumerate(items):
-            response.append({"item": item, "status": status[i].get("status") == "ok", "msg": status[i].get("msg")})
-
-        return web.json_response(data=response, status=web.HTTPOk.status_code, dumps=self.encoder.encode)
-
-    @route("GET", "api/logs", "logs")
-    async def logs(self, request: Request) -> Response:
-        """
-        Get recent logs
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        if not self.config.file_logging:
-            return web.json_response(
-                data={"error": "File logging is not enabled."}, status=web.HTTPNotFound.status_code
-            )
-
-        offset = int(request.query.get("offset", 0))
-        limit = int(request.query.get("limit", 100))
-        if limit < 1 or limit > 150:
-            limit = 50
-
-        logs_data = await read_logfile(
-            file=Path(self.config.config_path) / "logs" / "app.log",
-            offset=offset,
-            limit=limit,
-        )
-        return web.json_response(
-            data={
-                "logs": logs_data["logs"],
-                "offset": offset,
-                "limit": limit,
-                "next_offset": logs_data["next_offset"],
-                "end_is_reached": logs_data["end_is_reached"],
-            },
-            status=web.HTTPOk.status_code,
-            dumps=self.encoder.encode,
-        )
-
-    @route("GET", "api/conditions", "conditions")
-    async def conditions(self, _: Request) -> Response:
-        """
-        Get the conditions
-
-        Args:
-            _ (Request): The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        return web.json_response(
-            data=Conditions.get_instance().get_all(),
-            status=web.HTTPOk.status_code,
-            dumps=self.encoder.encode,
-        )
-
-    @route("PUT", "api/conditions", "conditions_add")
-    async def conditions_add(self, request: Request) -> Response:
-        """
-        Save Conditions.
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object
-
-        """
-        data = await request.json()
-
-        if not isinstance(data, list):
-            return web.json_response(
-                {"error": "Invalid request body expecting list with dicts."},
-                status=web.HTTPBadRequest.status_code,
-            )
-
-        items: list = []
-
-        cls = Conditions.get_instance()
-
-        for item in data:
-            if not isinstance(item, dict):
-                return web.json_response(
-                    {"error": "Invalid request body expecting list with dicts."},
-                    status=web.HTTPBadRequest.status_code,
-                )
-
-            if not item.get("name"):
-                return web.json_response(
-                    {"error": "name is required.", "data": item}, status=web.HTTPBadRequest.status_code
-                )
-
-            if not item.get("filter"):
-                return web.json_response(
-                    {"error": "filter is required.", "data": item}, status=web.HTTPBadRequest.status_code
-                )
-
-            if not item.get("cli"):
-                return web.json_response(
-                    {"error": "command options for yt-dlp is required.", "data": item},
-                    status=web.HTTPBadRequest.status_code,
-                )
-
-            if not item.get("id", None) or not validate_uuid(item.get("id"), version=4):
-                item["id"] = str(uuid.uuid4())
-
-            try:
-                cls.validate(item)
-            except ValueError as e:
-                return web.json_response(
-                    {"error": f"Failed to validate condition '{item.get('name')}'. '{e!s}'"},
-                    status=web.HTTPBadRequest.status_code,
-                )
-
-            items.append(init_class(Condition, item))
-        try:
-            items = cls.save(items=items).load().get_all()
-        except Exception as e:
-            LOG.exception(e)
-            return web.json_response(
-                {"error": "Failed to save conditions.", "message": str(e)},
-                status=web.HTTPInternalServerError.status_code,
-            )
-
-        await self._notify.emit(Events.CONDITIONS_UPDATE, data=items)
-        return web.json_response(data=items, status=web.HTTPOk.status_code, dumps=self.encoder.encode)
-
-    @route("POST", "api/conditions/test", "conditions_test")
-    async def conditions_test(self, request: Request) -> Response:
-        """
-        Test condition against URL.
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object
-
-        """
-        params = await request.json()
-
-        if not isinstance(params, dict):
-            return web.json_response(
-                {"error": "Invalid request body expecting dict."},
-                status=web.HTTPBadRequest.status_code,
-            )
-
-        url = params.get("url")
-        if not url:
-            return web.json_response({"error": "url is required."}, status=web.HTTPBadRequest.status_code)
-
-        cond = params.get("condition")
-        if not cond:
-            return web.json_response({"error": "condition is required."}, status=web.HTTPBadRequest.status_code)
-
-        try:
-            preset = params.get("preset", self.config.default_preset)
-            key = self.cache.hash(url + str(preset))
-            if not self.cache.has(key):
-                opts = {}
-                if ytdlp_proxy := self.config.get_ytdlp_args().get("proxy", None):
-                    opts["proxy"] = ytdlp_proxy
-                ytdlp_opts = YTDLPOpts.get_instance().preset(name=preset).add(opts).get_all()
-
-                data = extract_info(
-                    config=ytdlp_opts,
-                    url=url,
-                    debug=False,
-                    no_archive=True,
-                    follow_redirect=True,
-                    sanitize_info=True,
-                )
-                if not data:
-                    return web.json_response(
-                        data={"error": f"Failed to extract info from '{url!s}'."},
-                        status=web.HTTPBadRequest.status_code,
-                    )
-                self.cache.set(key=key, value=data, ttl=600)
-            else:
-                data = self.cache.get(key)
-        except Exception as e:
-            LOG.exception(e)
-            return web.json_response(
-                data={"error": f"Failed to extract video info. '{e!s}'"},
-                status=web.HTTPInternalServerError.status_code,
-            )
-
-        try:
-            from yt_dlp.utils import match_str
-
-            status = match_str(cond, data)
-        except Exception as e:
-            LOG.exception(e)
-            return web.json_response(
-                data={"error": str(e)},
-                status=web.HTTPBadRequest.status_code,
-            )
-
-        return web.json_response(
-            data={"status": status, "condition": cond, "data": data},
-            status=web.HTTPOk.status_code,
-            dumps=self.encoder.encode,
-        )
-
-    @route("GET", "api/presets", "presets")
-    async def presets(self, request: Request) -> Response:
-        """
-        Get the presets.
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        data = Presets.get_instance().get_all()
-        filter_fields = request.query.get("filter", None)
-
-        if filter_fields:
-            fields = [field.strip() for field in filter_fields.split(",")]
-            data = [{key: value for key, value in preset.__dict__.items() if key in fields} for preset in data]
-
-        return web.json_response(data=data, status=web.HTTPOk.status_code, dumps=self.encoder.encode)
-
-    @route("PUT", "api/presets", "presets_add")
-    async def presets_add(self, request: Request) -> Response:
-        """
-        Add presets.
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object
-
-        """
-        data = await request.json()
-
-        if not isinstance(data, list):
-            return web.json_response(
-                {"error": "Invalid request body expecting list with dicts."},
-                status=web.HTTPBadRequest.status_code,
-            )
-
-        presets: list = []
-
-        cls = Presets.get_instance()
-
-        for item in data:
-            if not isinstance(item, dict):
-                return web.json_response(
-                    {"error": "Invalid request body expecting list with dicts."},
-                    status=web.HTTPBadRequest.status_code,
-                )
-
-            if not item.get("name"):
-                return web.json_response(
-                    {"error": "name is required.", "data": item}, status=web.HTTPBadRequest.status_code
-                )
-
-            if not item.get("id", None) or not validate_uuid(item.get("id"), version=4):
-                item["id"] = str(uuid.uuid4())
-
-            try:
-                cls.validate(item)
-            except ValueError as e:
-                return web.json_response(
-                    {"error": f"Failed to validate preset '{item.get('name')}'. '{e!s}'"},
-                    status=web.HTTPBadRequest.status_code,
-                )
-
-            presets.append(Preset(**item))
-        try:
-            presets = cls.save(items=presets).load().get_all()
-        except Exception as e:
-            LOG.exception(e)
-            return web.json_response(
-                {"error": "Failed to save presets.", "message": str(e)},
-                status=web.HTTPInternalServerError.status_code,
-            )
-
-        await self._notify.emit(Events.PRESETS_UPDATE, data=presets)
-        return web.json_response(data=presets, status=web.HTTPOk.status_code, dumps=self.encoder.encode)
-
-    @route("GET", "api/tasks", "tasks")
-    async def tasks(self, _: Request) -> Response:
-        """
-        Get the tasks.
-
-        Args:
-            _: The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        return web.json_response(
-            data=Tasks.get_instance().get_all(), status=web.HTTPOk.status_code, dumps=self.encoder.encode
-        )
-
-    @route("PUT", "api/tasks", "tasks_add")
-    async def tasks_add(self, request: Request) -> Response:
-        """
-        Add tasks to the queue.
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object
-
-        """
-        data = await request.json()
-
-        if not isinstance(data, list):
-            return web.json_response(
-                {"error": "Invalid request body expecting list with dicts."},
-                status=web.HTTPBadRequest.status_code,
-            )
-
-        tasks: list = []
-
-        ins = Tasks.get_instance()
-
-        for item in data:
-            if not isinstance(item, dict):
-                return web.json_response(
-                    {"error": "Invalid request body expecting list with dicts."},
-                    status=web.HTTPBadRequest.status_code,
-                )
-
-            if not item.get("url"):
-                return web.json_response(
-                    {"error": "url is required.", "data": item}, status=web.HTTPBadRequest.status_code
-                )
-
-            if not item.get("id", None) or not validate_uuid(item.get("id"), version=4):
-                item["id"] = str(uuid.uuid4())
-
-            if not item.get("template", None):
-                item["template"] = ""
-
-            if not item.get("cli", None):
-                item["cli"] = ""
-
-            try:
-                Tasks.validate(item)
-            except ValueError as e:
-                return web.json_response(
-                    {"error": f"Failed to validate task '{item.get('name')}'. '{e!s}'"},
-                    status=web.HTTPBadRequest.status_code,
-                )
-
-            tasks.append(Task(**item))
-
-        try:
-            tasks = ins.save(tasks=tasks).load().get_all()
-        except Exception as e:
-            LOG.exception(e)
-            return web.json_response(
-                {"error": "Failed to save tasks.", "message": str(e)},
-                status=web.HTTPInternalServerError.status_code,
-            )
-
-        return web.json_response(data=tasks, status=web.HTTPOk.status_code, dumps=self.encoder.encode)
-
-    @route("DELETE", "api/history", "history_delete")
-    async def history_delete(self, request: Request) -> Response:
-        """
-        Delete an item from the queue.
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        data = await request.json()
-        ids = data.get("ids")
-        where = data.get("where")
-        if not ids or where not in ["queue", "done"]:
-            return web.json_response(
-                data={"error": "ids and where are required."}, status=web.HTTPBadRequest.status_code
-            )
-
-        remove_file: bool = bool(data.get("remove_file", True))
-
-        return web.json_response(
-            data=await (self.queue.cancel(ids) if where == "queue" else self.queue.clear(ids, remove_file=remove_file)),
-            status=web.HTTPOk.status_code,
-            dumps=self.encoder.encode,
-        )
-
-    @route("POST", "api/history/{id}", "history_item_update")
-    async def history_item_update(self, request: Request) -> Response:
-        """
-        Update an item in the history.
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        id: str = request.match_info.get("id")
-        if not id:
-            return web.json_response(data={"error": "id is required."}, status=web.HTTPBadRequest.status_code)
-
-        item = self.queue.done.get_by_id(id)
-        if not item:
-            return web.json_response(data={"error": "item not found."}, status=web.HTTPNotFound.status_code)
-
-        post = await request.json()
-        if not post:
-            return web.json_response(data={"error": "no data provided."}, status=web.HTTPBadRequest.status_code)
-
-        updated = False
-
-        for k, v in post.items():
-            if not hasattr(item.info, k):
+    def _load_modules(self, directory: Path):
+        package_name: str = str(directory.relative_to(Path(self.rootPath))).replace("/", ".")
+        LOG.debug(f"Loading modules from directory '{directory}' with package name '{package_name}'.")
+        for _, name, _ in pkgutil.iter_modules([directory]):
+            full_name: str = f"{package_name}.{name}"
+            if name.startswith("_"):
                 continue
-
-            if getattr(item.info, k) == v:
-                continue
-
-            updated = True
-            setattr(item.info, k, v)
-            LOG.debug(f"Updated '{k}' to '{v}' for '{item.info.id}'")
-
-        if updated:
-            self.queue.done.put(item)
-            await self._notify.emit(Events.UPDATE, data=item.info)
-
-        return web.json_response(
-            data=item.info,
-            status=web.HTTPOk.status_code if updated else web.HTTPNotModified.status_code,
-            dumps=self.encoder.encode,
-        )
-
-    @route("GET", "api/history", "history")
-    async def history(self, _: Request) -> Response:
-        """
-        Get the history.
-
-        Args:
-            _: The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        data: dict = {"queue": [], "history": []}
-        q = self.queue.get()
-
-        data["queue"].extend([q.get("queue", {}).get(k) for k in q.get("queue", {})])
-        data["history"].extend([q.get("done", {}).get(k) for k in q.get("done", {})])
-
-        return web.json_response(data=data, status=web.HTTPOk.status_code, dumps=self.encoder.encode)
-
-    @route("GET", "api/player/playlist/{file:.*}.m3u8", "playlist")
-    async def playlist(self, request: Request) -> Response:
-        """
-        Get the playlist.
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        file: str = request.match_info.get("file")
-
-        if not file:
-            return web.json_response(data={"error": "file is required"}, status=web.HTTPBadRequest.status_code)
-
-        base_path: str = self.config.base_path.rstrip("/")
-
-        try:
-            realFile, status = get_file(download_path=self.config.download_path, file=file)
-            if web.HTTPFound.status_code == status:
-                return Response(
-                    status=web.HTTPFound.status_code,
-                    headers={
-                        "Location": str(
-                            self.app.router["playlist"].url_for(
-                                file=str(realFile).replace(self.config.download_path, "").strip("/")
-                            )
-                        ),
-                    },
-                )
-
-            if web.HTTPNotFound.status_code == status:
-                return web.json_response(data={"error": f"File '{file}' does not exist."}, status=status)
-
-            return web.Response(
-                text=await Playlist(download_path=Path(self.config.download_path), url=f"{base_path}/").make(
-                    file=realFile
-                ),
-                headers={
-                    "Content-Type": "application/x-mpegURL",
-                    "Cache-Control": "no-cache",
-                    "Access-Control-Max-Age": "300",
-                },
-                status=web.HTTPOk.status_code,
-            )
-        except StreamingError as e:
-            return web.json_response(data={"error": str(e)}, status=web.HTTPNotFound.status_code)
-
-    @route("GET", "api/player/m3u8/{mode}/{file:.*}.m3u8", "m3u8")
-    async def m3u8(self, request: Request) -> Response:
-        """
-        Get the m3u8 file.
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        file: str = request.match_info.get("file")
-        mode: str = request.match_info.get("mode")
-
-        if mode not in ["video", "subtitle"]:
-            return web.json_response(
-                data={"error": "Only video and subtitle modes are supported."}, status=web.HTTPBadRequest.status_code
-            )
-
-        if not file:
-            return web.json_response(data={"error": "file is required"}, status=web.HTTPBadRequest.status_code)
-
-        duration = request.query.get("duration", None)
-
-        if "subtitle" in mode:
-            if not duration:
-                return web.json_response(data={"error": "duration is required."}, status=web.HTTPBadRequest.status_code)
-
-            duration = float(duration)
-
-        base_path: str = self.config.base_path.rstrip("/")
-
-        try:
-            cls = M3u8(download_path=Path(self.config.download_path), url=f"{base_path}/")
-
-            realFile, status = get_file(download_path=self.config.download_path, file=file)
-            if web.HTTPFound.status_code == status:
-                return Response(
-                    status=status,
-                    headers={
-                        "Location": str(
-                            self.app.router["m3u8"].url_for(
-                                mode=mode,
-                                file=str(realFile).replace(self.config.download_path, "").strip("/"),
-                            )
-                        ),
-                    },
-                )
-
-            if web.HTTPNotFound.status_code == status:
-                return web.json_response(data={"error": f"File '{file}' does not exist."}, status=status)
-
-            if "subtitle" in mode:
-                text = await cls.make_subtitle(file=realFile, duration=duration)
-            else:
-                text = await cls.make_stream(file=realFile)
-        except StreamingError as e:
-            LOG.exception(e)
-            return web.json_response(data={"error": str(e)}, status=web.HTTPNotFound.status_code)
-
-        return web.Response(
-            text=text,
-            headers={
-                "Content-Type": "application/x-mpegURL",
-                "Cache-Control": "no-cache",
-                "Access-Control-Max-Age": "300",
-            },
-            status=web.HTTPOk.status_code,
-        )
-
-    @route("GET", r"api/player/segments/{segment:\d+}/{file:.*}.ts", "segments")
-    async def segments(self, request: Request) -> Response:
-        """
-        Get the segments.
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        file: str = request.match_info.get("file")
-        segment: int = request.match_info.get("segment")
-        sd: int = request.query.get("sd")
-        vc: int = int(request.query.get("vc", 0))
-        ac: int = int(request.query.get("ac", 0))
-
-        if not file:
-            return web.json_response(data={"error": "file is required"}, status=web.HTTPBadRequest.status_code)
-
-        if not segment:
-            return web.json_response(data={"error": "segment id is required."}, status=web.HTTPBadRequest.status_code)
-
-        realFile, status = get_file(download_path=self.config.download_path, file=file)
-        if web.HTTPFound.status_code == status:
-            return Response(
-                status=status,
-                headers={
-                    "Location": str(
-                        self.app.router["segments"].url_for(
-                            segment=segment,
-                            file=str(realFile).replace(self.config.download_path, "").strip("/"),
-                        )
-                    ),
-                },
-            )
-
-        if web.HTTPNotFound.status_code == status:
-            return web.json_response(data={"error": f"File '{file}' does not exist."}, status=status)
-
-        mtime = realFile.stat().st_mtime
-
-        if request.if_modified_since and request.if_modified_since.timestamp() == mtime:
-            lastMod = time.strftime("%a, %d %b %Y %H:%M:%S GMT", datetime.fromtimestamp(mtime, tz=UTC).timetuple())
-            return web.Response(status=web.HTTPNotModified.status_code, headers={"Last-Modified": lastMod})
-
-        resp = web.StreamResponse(
-            status=web.HTTPOk.status_code,
-            headers={
-                "Content-Type": "video/mpegts",
-                "X-Accel-Buffering": "no",
-                "Access-Control-Allow-Origin": "*",
-                "Pragma": "public",
-                "Cache-Control": f"public, max-age={time.time() + 31536000}",
-                "Last-Modified": time.strftime(
-                    "%a, %d %b %Y %H:%M:%S GMT", datetime.fromtimestamp(mtime, tz=UTC).timetuple()
-                ),
-                "Expires": time.strftime(
-                    "%a, %d %b %Y %H:%M:%S GMT", datetime.fromtimestamp(time.time() + 31536000, tz=UTC).timetuple()
-                ),
-            },
-        )
-
-        await resp.prepare(request)
-
-        await Segments(
-            download_path=self.config.download_path,
-            index=int(segment),
-            duration=float(f"{float(sd if sd else M3u8.duration):.6f}"),
-            vconvert=vc == 1,
-            aconvert=ac == 1,
-        ).stream(realFile, resp)
-
-        return resp
-
-    @route("GET", "api/player/subtitle/{file:.*}.vtt", "subtitles")
-    async def subtitles(self, request: Request) -> Response:
-        """
-        Get the subtitles.
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        file: str = request.match_info.get("file")
-
-        if not file:
-            return web.json_response(data={"error": "file is required"}, status=web.HTTPBadRequest.status_code)
-
-        realFile, status = get_file(download_path=self.config.download_path, file=file)
-        if web.HTTPFound.status_code == status:
-            return Response(
-                status=status,
-                headers={
-                    "Location": str(
-                        self.app.router["subtitles"].url_for(
-                            file=str(realFile).replace(self.config.download_path, "").strip("/")
-                        )
-                    ),
-                },
-            )
-
-        if web.HTTPNotFound.status_code == status:
-            return web.json_response(data={"error": f"File '{file}' does not exist."}, status=status)
-
-        mtime = realFile.stat().st_mtime
-
-        if request.if_modified_since and request.if_modified_since.timestamp() == mtime:
-            lastMod = time.strftime("%a, %d %b %Y %H:%M:%S GMT", datetime.fromtimestamp(mtime, tz=UTC).timetuple())
-            return web.Response(status=web.HTTPNotModified.status_code, headers={"Last-Modified": lastMod})
-
-        return web.Response(
-            body=await Subtitle().make(file=realFile),
-            headers={
-                "Content-Type": "text/vtt; charset=UTF-8",
-                "X-Accel-Buffering": "no",
-                "Access-Control-Allow-Origin": "*",
-                "Pragma": "public",
-                "Cache-Control": f"public, max-age={time.time() + 31536000}",
-                "Last-Modified": time.strftime(
-                    "%a, %d %b %Y %H:%M:%S GMT", datetime.fromtimestamp(mtime, tz=UTC).timetuple()
-                ),
-                "Expires": time.strftime(
-                    "%a, %d %b %Y %H:%M:%S GMT", datetime.fromtimestamp(time.time() + 31536000, tz=UTC).timetuple()
-                ),
-            },
-            status=web.HTTPOk.status_code,
-        )
-
-    @route("GET", "/", "index")
-    async def index(self, _: Request) -> Response:
-        """
-        Get the index file.
-
-        Args:
-            _: The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        if "/index.html" not in self._static_holder:
-            LOG.error("Static frontend files not found.")
-            return web.json_response(
-                data={"error": "File not found.", "file": "/index.html"}, status=web.HTTPNotFound.status_code
-            )
-
-        data = self._static_holder["/index.html"]
-        return web.Response(
-            body=data.get("content"),
-            content_type=data.get("content_type"),
-            charset="utf-8",
-            status=web.HTTPOk.status_code,
-        )
-
-    @route("GET", "api/thumbnail", "get_thumbnail")
-    async def get_thumbnail(self, request: Request) -> Response:
-        """
-        Get the thumbnail.
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        url = request.query.get("url")
-        if not url:
-            return web.json_response(data={"error": "URL is required."}, status=web.HTTPForbidden.status_code)
-
-        try:
-            validate_url(url)
-        except ValueError as e:
-            return web.json_response(data={"error": str(e)}, status=web.HTTPForbidden.status_code)
-
-        try:
-            ytdlp_args = self.config.get_ytdlp_args()
-            opts = {
-                "proxy": ytdlp_args.get("proxy", None),
-                "headers": {
-                    "User-Agent": ytdlp_args.get(
-                        "user_agent", request.headers.get("User-Agent", f"YTPTube/{self.config.version}")
-                    ),
-                },
-            }
-
-            async with httpx.AsyncClient(**opts) as client:
-                LOG.debug(f"Fetching thumbnail from '{url}'.")
-                response = await client.request(method="GET", url=url)
-                return web.Response(
-                    body=response.content,
-                    headers={
-                        "Content-Type": response.headers.get("Content-Type"),
-                        "Pragma": "public",
-                        "Access-Control-Allow-Origin": "*",
-                        "Cache-Control": f"public, max-age={time.time() + 31536000}",
-                        "Expires": time.strftime(
-                            "%a, %d %b %Y %H:%M:%S GMT",
-                            datetime.fromtimestamp(time.time() + 31536000, tz=UTC).timetuple(),
-                        ),
-                    },
-                )
-        except Exception as e:
-            LOG.error(f"Error fetching thumbnail from '{url}'. '{e}'.")
-            return web.json_response(
-                data={"error": "failed to retrieve the thumbnail."}, status=web.HTTPInternalServerError.status_code
-            )
-
-    @route("GET", "api/random/background", "get_background")
-    async def get_background(self, request: Request) -> Response:
-        """
-        Get random background.
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        backend = None
-
-        if self.isRequestingBackground:
-            return web.Response(status=web.HTTPTooManyRequests.status_code)
-
-        try:
-            self.isRequestingBackground = True
-            backend = random.choice(self.config.pictures_backends)  # noqa: S311
-            CACHE_KEY_BING = "random_background_bing"
-            CACHE_KEY = "random_background"
-
-            if self.cache.has(CACHE_KEY) and not request.query.get("force", False):
-                data = await self.cache.aget(CACHE_KEY)
-                return web.Response(
-                    body=data.get("content"),
-                    headers={
-                        "X-Cache": "HIT",
-                        "X-Cache-TTL": str(await self.cache.attl(CACHE_KEY)),
-                        "X-Image-Via": data.get("backend"),
-                        **data.get("headers"),
-                    },
-                )
-
-            ytdlp_args = self.config.get_ytdlp_args()
-            opts = {
-                "proxy": ytdlp_args.get("proxy", None),
-                "headers": {
-                    "User-Agent": ytdlp_args.get("user_agent", f"YTPTube/{self.config.version}"),
-                },
-            }
-
-            async with httpx.AsyncClient(**opts) as client:
-                if backend.startswith("https://www.bing.com/HPImageArchive.aspx"):
-                    if not self.cache.has(CACHE_KEY_BING):
-                        response = await client.request(method="GET", url=backend)
-                        if response.status_code != web.HTTPOk.status_code:
-                            return web.json_response(
-                                data={"error": "failed to retrieve the random background image."},
-                                status=web.HTTPInternalServerError.status_code,
-                            )
-
-                        img_url: str | None = ag(response.json(), "images.0.url")
-                        if not img_url:
-                            return web.json_response(
-                                data={"error": "failed to retrieve the random background image."},
-                                status=web.HTTPInternalServerError.status_code,
-                            )
-
-                        backend = f"https://www.bing.com{img_url}"
-                        await self.cache.aset(key=CACHE_KEY_BING, value=backend, ttl=3600 * 24)
-                    else:
-                        backend: str = await self.cache.aget(CACHE_KEY_BING)
-
-                LOG.debug(f"Requesting random picture from '{backend!s}'.")
-
-                response = await client.request(method="GET", url=backend, follow_redirects=True)
-
-                if response.status_code != web.HTTPOk.status_code:
-                    return web.json_response(
-                        data={"error": "failed to retrieve the random background image."},
-                        status=web.HTTPInternalServerError.status_code,
-                    )
-
-                data = {
-                    "content": response.content,
-                    "backend": urlparse(backend).netloc,
-                    "headers": {
-                        "Content-Type": response.headers.get("Content-Type", "image/jpeg"),
-                        "Content-Length": str(len(response.content)),
-                    },
-                }
-
-                await self.cache.aset(key=CACHE_KEY, value=data, ttl=3600)
-
-                LOG.debug(f"Random background image from '{backend!s}' cached.")
-
-                return web.Response(
-                    body=data.get("content"),
-                    headers={
-                        "X-Cache": "MISS",
-                        "X-Cache-TTL": "3600",
-                        "X-Image-Via": data.get("backend"),
-                        **data.get("headers"),
-                    },
-                )
-        except Exception as e:
-            LOG.error(f"Failed to request random background image from '{backend!s}'.'. '{e!s}'.")
-            return web.json_response(
-                data={"error": "failed to retrieve the random background image."},
-                status=web.HTTPInternalServerError.status_code,
-            )
-        finally:
-            self.isRequestingBackground = False
-
-    @route("GET", "api/file/ffprobe/{file:.*}", "ffprobe")
-    async def get_ffprobe(self, request: Request) -> Response:
-        """
-        Get the ffprobe data.
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        file: str = request.match_info.get("file")
-        if not file:
-            return web.json_response(data={"error": "file is required."}, status=web.HTTPBadRequest.status_code)
-
-        try:
-            realFile, status = get_file(download_path=self.config.download_path, file=file)
-            if web.HTTPFound.status_code == status:
-                return Response(
-                    status=web.HTTPFound.status_code,
-                    headers={
-                        "Location": str(
-                            self.app.router["ffprobe"].url_for(
-                                file=str(realFile.relative_to(self.config.download_path).as_posix()).strip("/")
-                            )
-                        ),
-                    },
-                )
-
-            if web.HTTPNotFound.status_code == status:
-                return web.json_response(data={"error": f"File '{file}' does not exist."}, status=status)
-
-            return web.json_response(
-                data=await ffprobe(realFile), status=web.HTTPOk.status_code, dumps=self.encoder.encode
-            )
-        except Exception as e:
-            return web.json_response(data={"error": str(e)}, status=web.HTTPInternalServerError.status_code)
-
-    @route("GET", "api/file/info/{file:.*}", "file_info")
-    async def get_file_info(self, request: Request) -> Response:
-        """
-        Get file info
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        file: str = request.match_info.get("file")
-        if not file:
-            return web.json_response(data={"error": "file is required."}, status=web.HTTPBadRequest.status_code)
-
-        try:
-            realFile, status = get_file(download_path=self.config.download_path, file=file)
-            if web.HTTPFound.status_code == status:
-                return Response(
-                    status=web.HTTPFound.status_code,
-                    headers={
-                        "Location": str(
-                            self.app.router["file_info"].url_for(
-                                file=str(realFile.relative_to(self.config.download_path).as_posix()).strip("/")
-                            )
-                        ),
-                    },
-                )
-
-            if web.HTTPNotFound.status_code == status:
-                return web.json_response(data={"error": f"File '{file}' does not exist."}, status=status)
-
-            ff_info = await ffprobe(realFile)
-
-            response = {
-                "title": realFile.stem,
-                "ffprobe": ff_info,
-                "mimetype": get_mime_type(ff_info.get("metadata", {}), realFile),
-                "sidecar": get_file_sidecar(realFile),
-            }
-
-            for key in response["sidecar"]:
-                for i, f in enumerate(response["sidecar"][key]):
-                    response["sidecar"][key][i]["file"] = str(
-                        realFile.with_name(f["file"].name).relative_to(self.config.download_path)
-                    ).strip("/")
-
-            return web.json_response(data=response, status=web.HTTPOk.status_code, dumps=self.encoder.encode)
-        except Exception as e:
-            LOG.exception(e)
-            return web.json_response(data={"error": str(e)}, status=web.HTTPInternalServerError.status_code)
-
-    @route("GET", "api/notifications", "notifications")
-    async def notifications(self, _: Request) -> Response:
-        """
-        Get the notification targets.
-
-        Args:
-            _: The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        return web.json_response(
-            data={
-                "notifications": Notification.get_instance().get_targets(),
-                "allowedTypes": list(NotificationEvents.get_events().values()),
-            },
-            status=web.HTTPOk.status_code,
-            dumps=self.encoder.encode,
-        )
-
-    @route("PUT", "api/notifications", "notification_add")
-    async def notification_add(self, request: Request) -> Response:
-        """
-        Add notification targets.
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        post = await request.json()
-        if not isinstance(post, list):
-            return web.json_response(
-                {"error": "Invalid request body expecting list with dicts."},
-                status=web.HTTPBadRequest.status_code,
-            )
-
-        targets: list = []
-
-        ins = Notification.get_instance()
-        for item in post:
-            if not isinstance(item, dict):
-                return web.json_response(
-                    {"error": "Invalid request body expecting list with dicts."},
-                    status=web.HTTPBadRequest.status_code,
-                )
-
-            if not item.get("id", None) or validate_uuid(item.get("id"), version=4):
-                item["id"] = str(uuid.uuid4())
-
             try:
-                Notification.validate(item)
-            except ValueError as e:
-                return web.json_response(
-                    {"error": f"Invalid notification target settings. {e!s}", "data": item},
-                    status=web.HTTPBadRequest.status_code,
-                )
-
-            targets.append(ins.make_target(item))
-
-        try:
-            ins.save(targets=targets)
-            ins.load()
-        except Exception as e:
-            LOG.exception(e)
-            return web.json_response({"error": "Failed to save tasks."}, status=web.HTTPInternalServerError.status_code)
-
-        data = {"notifications": targets, "allowedTypes": list(NotificationEvents.get_events().values())}
-
-        return web.json_response(data=data, status=web.HTTPOk.status_code, dumps=self.encoder.encode)
-
-    @route("POST", "api/notifications/test", "notification_test")
-    async def notification_test(self, _: Request) -> Response:
-        """
-        Test the notification.
-
-        Args:
-            _: The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        data = message("test", "This is a test notification.")
-
-        await self._notify.emit(Events.TEST, data=data)
-
-        return web.json_response(data=data, status=web.HTTPOk.status_code, dumps=self.encoder.encode)
-
-    @route("GET", "api/file/browser/{path:.*}", "file_browser")
-    async def file_browser(self, request: Request) -> Response:
-        """
-        Get the file browser.
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        if not self.config.browser_enabled:
-            return web.json_response(data={"error": "File browser is disabled."}, status=web.HTTPForbidden.status_code)
-
-        req_path: str = request.match_info.get("path")
-        req_path: str = "/" if not req_path else unquote_plus(req_path)
-
-        test: Path = Path(self.config.download_path).joinpath(req_path)
-        if not test.exists():
-            return web.json_response(
-                data={"error": f"path '{req_path}' does not exist."}, status=web.HTTPNotFound.status_code
-            )
-
-        if not test.is_dir() and not test.is_symlink():
-            return web.json_response(
-                data={"error": f"path '{req_path}' is not a directory."}, status=web.HTTPBadRequest.status_code
-            )
-
-        try:
-            return web.json_response(
-                data={
-                    "path": req_path,
-                    "contents": get_files(base_path=Path(self.config.download_path), dir=req_path),
-                },
-                status=web.HTTPOk.status_code,
-                dumps=self.encoder.encode,
-            )
-        except OSError as e:
-            LOG.exception(e)
-            return web.json_response(data={"error": str(e)}, status=web.HTTPInternalServerError.status_code)
-
-    @route("GET", "api/dev/loop")
-    async def debug_asyncio(self, _: Request) -> Response:
-        if not self.config.is_dev():
-            return web.json_response(
-                data={"error": "This endpoint is only available in development mode."},
-                status=web.HTTPForbidden.status_code,
-            )
-
-        import traceback
-
-        tasks = []
-        for task in asyncio.all_tasks():
-            task_info = {"task": str(task), "stack": []}
-            for frame in task.get_stack():
-                formatted = traceback.format_stack(f=frame)
-                task_info["stack"].extend(formatted)
-
-            tasks.append(task_info)
-
-        return web.json_response(
-            data={
-                "total_tasks": len(tasks),
-                "loop": str(asyncio.get_event_loop()),
-                "tasks": tasks,
-            },
-            status=web.HTTPOk.status_code,
-            dumps=self.encoder.encode,
-        )
-
-    @route("GET", "api/dev/workers", "pool_list")
-    async def pool_list(self, _) -> Response:
-        """
-        Get the workers status.
-
-        Args:
-            _: The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        if not self.config.is_dev():
-            return web.json_response(
-                {"error": "This endpoint is only available in development mode."},
-                status=web.HTTPNotFound.status_code,
-            )
-
-        if self.queue.pool is None:
-            return web.json_response({"error": "Worker pool not initialized."}, status=web.HTTPNotFound.status_code)
-
-        status = self.queue.pool.get_workers_status()
-
-        data = []
-
-        for worker in status:
-            worker_status = status.get(worker)
-            data.append(
-                {
-                    "id": worker,
-                    "data": {"status": "Waiting for download."} if worker_status is None else worker_status,
-                }
-            )
-
-        return web.json_response(
-            data={
-                "open": self.queue.pool.has_open_workers(),
-                "count": self.queue.pool.get_available_workers(),
-                "workers": data,
-            },
-            status=web.HTTPOk.status_code,
-            dumps=lambda obj: json.dumps(obj, default=lambda o: f"<<non-serializable: {type(o).__qualname__}>>"),
-        )
-
-    @route("POST", "api/dev/workers", "pool_start")
-    async def pool_restart(self, _) -> Response:
-        """
-        Restart the workers pool.
-
-        Args:
-            _: The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        if not self.config.is_dev():
-            return web.json_response(
-                {"error": "This endpoint is only available in development mode."},
-                status=web.HTTPNotFound.status_code,
-            )
-
-        if self.queue.pool is None:
-            return web.json_response({"error": "Worker pool not initialized."}, status=web.HTTPNotFound.status_code)
-
-        self.queue.pool.start()
-
-        return web.json_response({"message": "Workers pool being restarted."}, status=web.HTTPOk.status_code)
-
-    @route("PATCH", "api/dev/workers/{id}", "worker_restart")
-    async def worker_restart(self, request: Request) -> Response:
-        """
-        Restart a worker.
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object
-
-        """
-        if not self.config.is_dev():
-            return web.json_response(
-                {"error": "This endpoint is only available in development mode."},
-                status=web.HTTPNotFound.status_code,
-            )
-
-        worker_id: str = request.match_info.get("id")
-        if not worker_id:
-            return web.json_response({"error": "worker id is required."}, status=web.HTTPBadRequest.status_code)
-
-        if self.queue.pool is None:
-            return web.json_response({"error": "Worker pool not initialized."}, status=web.HTTPNotFound.status_code)
-
-        status = await self.queue.pool.restart(worker_id, "requested by user.")
-
-        return web.json_response({"status": "restarted" if status else "in_error_state"}, status=web.HTTPOk.status_code)
-
-    @route("DELETE", "api/dev/workers/{id}", "worker_stop")
-    async def worker_stop(self, request: Request) -> Response:
-        """
-        Stop a worker.
-
-        Args:
-            request (Request): The request object.
-
-        Returns:
-            Response: The response object.
-
-        """
-        if not self.config.is_dev():
-            return web.json_response(
-                {"error": "This endpoint is only available in development mode."},
-                status=web.HTTPNotFound.status_code,
-            )
-
-        worker_id: str = request.match_info.get("id")
-        if not worker_id:
-            raise web.HTTPBadRequest(text="worker id is required.")
-
-        if self.queue.pool is None:
-            return web.json_response({"error": "Worker pool not initialized."}, status=web.HTTPNotFound.status_code)
-
-        status = await self.queue.pool.stop(worker_id, "requested by user.")
-
-        return web.json_response({"status": "stopped" if status else "in_error_state"}, status=web.HTTPOk.status_code)
+                LOG.debug(f"Loading module '{full_name}'.")
+                importlib.import_module(full_name)
+            except ImportError as e:
+                LOG.error(f"Failed to import module '{full_name}': {e}")
