+++ app/native.py
@@ -1,9 +1,17 @@
+#!/usr/bin/env python3
+import sys
+from pathlib import Path
+
+APP_ROOT = str((Path(__file__).parent / "..").resolve())
+if APP_ROOT not in sys.path:
+    sys.path.insert(0, APP_ROOT)
+
+
 import json
 import os
 import queue
 import socket
 import threading
-from pathlib import Path
 
 ready = threading.Event()
 exception_holder = queue.Queue()
